requirements.txt
Django==1.11
django-admin-rangefilter==0.3.2
django-filter==1.1.0
django-model-utils==3.1.1
djangorestframework==3.6.2
psycopg2==2.7.4
pytz==2018.3